Move files to Receive folder
