+++ app/components/UI/PaymentRequest/index.js
@@ -60,7 +60,7 @@ import { selectTokens } from '../../../selectors/tokensController';
 import { selectContractExchangeRates } from '../../../selectors/tokenRatesController';
 import { selectSelectedInternalAccountChecksummedAddress } from '../../../selectors/accountsController';
 
-import { RequestPaymentViewSelectors } from '../../../../e2e/selectors/RequestPaymentView.selectors';
+import { RequestPaymentViewSelectors } from '../../../../e2e/selectors/Receive/RequestPaymentView.selectors';
 
 const KEYBOARD_OFFSET = 120;
 const createStyles = (colors) =>
