+++ e2e/selectors/Receive/RequestPaymentView.selectors.js
