+++ e2e/selectors/Receive/SendLinkView.selectors.js
